refactor: type 선언 파일 통합
