+++ src/apis/schedule/getDailyWorkers.ts
@@ -1,25 +1,14 @@
 import instance from 'apis/instance';
+import { TimeWorkerListData } from 'apis/types';
 
 export const getDailyWorkers = (params: GetDailyWorkersRequest): Promise<GetDailyWorkerResponse> => {
   return instance.get(`/schedule/fix/day`, { params });
 };
-
-export interface DailyWorkTimeData {
-  title: string;
-  startTime: string;
-  endTime: string;
-  workerList: WorkerData[];
-}
-
-export interface WorkerData {
-  memberId: number;
-  name: string;
-}
 
 interface GetDailyWorkersRequest {
   date: string;
 }
 
 interface GetDailyWorkerResponse {
-  data: { schedule: DailyWorkTimeData[] };
+  data: { schedule: TimeWorkerListData[] };
 }
